+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/client/service/sorting/AlphabeticallySortingService.java
@@ -1,6 +1,7 @@
 package com.nemonotfound.nemos.inventory.sorting.client.service.sorting;
 
 import com.nemonotfound.nemos.inventory.sorting.client.model.SlotItem;
+import com.nemonotfound.nemos.inventory.sorting.client.service.TooltipService;
 import com.nemonotfound.nemos.inventory.sorting.client.service.SlotSwappingService;
 import net.minecraft.client.Minecraft;
 
@@ -10,13 +11,13 @@ public class AlphabeticallySortingService extends AbstractSortingService {
 
     private static AlphabeticallySortingService INSTANCE;
 
-    private AlphabeticallySortingService(SlotSwappingService inventorySwapService, Minecraft minecraft) {
-        super(inventorySwapService, minecraft);
+    private AlphabeticallySortingService(SlotSwappingService inventorySwapService, TooltipService tooltipService, Minecraft minecraft) {
+        super(inventorySwapService, tooltipService, minecraft);
     }
 
     public static AlphabeticallySortingService getInstance() {
         if(INSTANCE == null) {
-            INSTANCE = new AlphabeticallySortingService(SlotSwappingService.getInstance(), Minecraft.getInstance());
+            INSTANCE = new AlphabeticallySortingService(SlotSwappingService.getInstance(), TooltipService.getInstance(), Minecraft.getInstance());
         }
 
         return INSTANCE;
@@ -24,10 +25,6 @@ public class AlphabeticallySortingService extends AbstractSortingService {
 
     @Override
     protected Comparator<SlotItem> comparator() {
-        return Comparator.comparing(
-                slotItem -> slotItem.itemStack()
-                        .getItemName()
-                        .getString()
-        );
+        return comparatorByName();
     }
 }