+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Changes
 - Internal changes & cleanup (Slightly faster sorting process)
+- Items are now sorted by enchantment names as well
 
 ## Fixes
 - Search box won't lose focus by clicking outside